+++ src/models/nematus.h
@@ -26,11 +26,9 @@ public:
   }
 
   void load(Ptr<ExpressionGraph> graph,
-            const std::string& name,
+            const std::vector<io::Item>& items,
             bool /*markReloaded*/ = true) override {
-    LOG(info, "Loading model from {}", name);
-    // load items from .npz file
-    auto ioItems = io::loadItems(name);
+    auto ioItems = items;
     // map names and remove a dummy matrix 'decoder_c_tt' from items to avoid creating isolated node
     for(auto it = ioItems.begin(); it != ioItems.end();) {
       // for backwards compatibility, turn one-dimensional vector into two dimensional matrix with first dimension being 1 and second dimension of the original size
@@ -41,7 +39,7 @@ public:
         it->shape.set(0, 1);
         it->shape.set(1, dim);
       }
-      
+
       if(it->name == "decoder_c_tt") {
         it = ioItems.erase(it);
       } else if(it->name == "uidx") {
@@ -57,6 +55,14 @@ public:
     }
     // load items into the graph
     graph->load(ioItems);
+  }
+
+  void load(Ptr<ExpressionGraph> graph,
+            const std::string& name,
+            bool /*markReloaded*/ = true) override {
+    LOG(info, "Loading model from {}", name);
+    auto ioItems = io::loadItems(name);
+    load(graph, ioItems);
   }
 
   void save(Ptr<ExpressionGraph> graph,